+++ components/analysis/index.ts
@@ -1,0 +1,1 @@
+export { default as AnalysisTool } from './AnalysisTool'
