+++ components/dialogs/ManualEntryDialog.tsx
