+++ components/search/index.ts
@@ -8,4 +8,3 @@ export { ExpressionPreview } from "./ExpressionPreview"
 export { ExpressionLegend } from "./ExpressionLegend"
 export { SavedConditionsList } from "./SavedConditionsList"
 export { SaveConditionDialog } from "./SaveConditionDialog"
-export { TriggerSignalDialog } from "./TriggerSignalDialog"
